previous-poc/previous_poc/quiet-python-poc-3/requirements.txt
PyYAML
rich
textual
pytest
typing-extensions
PyNaCl
